requirements.txt
requests~=2.32.3
torch~=2.7.0
pillow~=11.2.1
torchvision~=0.22.0
matplotlib~=3.10.3
tqdm~=4.67.1
PyYAML~=6.0.2